Separate Pancake libraries from C Microkit

+++ build_sdk.py
@@ -627,18 +627,25 @@ def build_sel4(
     shutil.copy(invocations_all, dest)
     dest.chmod(0o744)
 
-    include_dir = root_dir / "board" / board.name / config.name / "include"
-    for source in ("kernel_Config", "libsel4", "libsel4/sel4_Config", "libsel4/autoconf"):
+    def copy_kernel_includes(target_dir, source_dir):
         source_dir = sel4_install_dir / source / "include"
         for p in source_dir.rglob("*"):
             if not p.is_file():
                 continue
             rel = p.relative_to(source_dir)
-            dest = include_dir / rel
+            dest = target_dir / rel
             dest.parent.mkdir(exist_ok=True, parents=True)
             dest.unlink(missing_ok=True)
             shutil.copy(p, dest)
             dest.chmod(0o744)
+
+    for source in ("kernel_Config", "libsel4", "libsel4/sel4_Config", "libsel4/autoconf"):
+        source_dir = sel4_install_dir / source / "include"
+        copy_kernel_includes(root_dir / "board" / board.name / config.name / "include", source_dir)
+        copy_kernel_includes(
+            root_dir / "board" / board.name / config.name / "include" / "pancake-c-interop",
+            source_dir
+        )
 
     if not board.arch.is_x86():
         # only non-x86 platforms have this file to describe memory regions
@@ -707,7 +714,7 @@ def build_lib_component(
 ) -> None:
     """Build a specific library component.
 
-    Right now this is just libmicrokit.a
+    Right now these are just libmicrokit.a and libmicrokitpnk.a
     """
     sel4_dir = root_dir / "board" / board.name / config.name
     build_dir = build_dir / board.name / config.name / component_name
@@ -736,10 +743,11 @@ def build_lib_component(
 
     link_script = Path(component_name) / "microkit.ld"
     dest = lib_dir / "microkit.ld"
-    dest.unlink(missing_ok=True)
-    shutil.copy(link_script, dest)
-    # Make output read-only
-    dest.chmod(0o744)
+    if link_script.exists():
+        dest.unlink(missing_ok=True)
+        shutil.copy(link_script, dest)
+        # Make output read-only
+        dest.chmod(0o744)
 
     include_dir = root_dir / "board" / board.name / config.name / "include"
     source_dir = Path(component_name) / "include"
@@ -747,8 +755,10 @@ def build_lib_component(
         if not p.is_file():
             continue
         rel = p.relative_to(source_dir)
-        if rel.parts[0] == "pancake" and not board.arch.is_riscv():
-            # The Pancake components currently only support RISC-V
+        # The Pancake components currently only support RISC-V
+        if not board.arch.is_riscv() and rel.parts[0] in ("panmicrokit", "pansel4"):
+            continue
+        if not board.arch.is_riscv() and rel.name == "panmicrokit.h":
             continue
         dest = include_dir / rel
         dest.parent.mkdir(exist_ok=True, parents=True)
@@ -931,6 +941,9 @@ def main() -> None:
 
                 build_elf_component("monitor", root_dir, build_dir, board, config, args.llvm, [])
                 build_lib_component("libmicrokit", root_dir, build_dir, board, config, args.llvm)
+                if board.arch.is_riscv():
+                    # the Pancake version is RISCV64 only for now
+                    build_lib_component("libmicrokitpnk", root_dir, build_dir, board, config, args.llvm)
                 if not args.skip_initialiser:
                     build_initialiser("initialiser", root_dir, build_dir, board, config)
